Course-1-PyTorch-Exploration/Module 2/Lab_1_mnist_classifier/requirements.txt
numpy==2.3.2
torch==2.6.0
torchvision==0.21.0
matplotlib==3.9.0
pillow==11.3.0